Handroll type checking for adhoc actions
I want to allow users to register adhoc with actions as one of: [, 'method'], closure or object with __invoke, so I can't simply require an object that implements an interface.
Luckily, validation runs only once: at registration, so it won't be too much of a burden to check the types through reflection.